+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Thanks to the following contributors who worked on this release:
   - Updated French translation (#406)
 
 ### Changed
+- Chord diagrams now display the fret number of the top fret instead of the next fret after it (#408)
 - Updated the minimum required version of `RtMidi` to 4.0
 
 ### Fixed
@@ -24,6 +25,7 @@ Thanks to the following contributors who worked on this release:
 - Fixed a potential crash when importing Guitar Pro files (#442)
 - Fixed a bug where the scroll bar's range did not update when the visible height of the score shrank (#443)
 - Fixed a bug where the caret was rendered when printing or displaying a print preview (#441)
+- Fixed issues where the top fret number of a chord diagram could be cut off in the chord name dialog (#408)
 
 ## [Alpha 19] - 2022-10-17